requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
requests==2.31.0
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7
shopify-python-api==1.0.1
stripe==5.5.0
paypalrestsdk==1.13.1
pandas==2.0.3
openpyxl==3.1.2
APScheduler==3.10.4
PyJWT==2.8.0